chore: move signed document to models, signed document manager to repositories
