+++ catalyst_voices/packages/internal/catalyst_voices_models/lib/src/document/data/document_data.dart
@@ -1,5 +1,4 @@
 import 'package:catalyst_voices_models/catalyst_voices_models.dart';
-import 'package:catalyst_voices_shared/catalyst_voices_shared.dart';
 import 'package:equatable/equatable.dart';
 
 /// Represents raw data about any document. It may be published signed document
@@ -10,7 +9,7 @@ import 'package:equatable/equatable.dart';
 /// For example one [Document] of [DocumentType.proposalDocument] require
 /// document itself and template which is different document.
 ///
-/// [DocumentData] can be created from [SignedDocument] which comes from
+/// [DocumentData] can be created from SignedDocument which comes from
 /// backend or locally as work in progress.
 final class DocumentData extends Equatable {
   final DocumentDataMetadata metadata;
@@ -21,12 +20,12 @@ final class DocumentData extends Equatable {
     required this.content,
   });
 
-  /// Syntax sugar. Should use [DocumentDataMetadata.selfRef].
-  DocumentRef get ref => metadata.selfRef;
-
   @override
   List<Object?> get props => [
         metadata,
         content,
       ];
+
+  /// Syntax sugar. Should use [DocumentDataMetadata.selfRef].
+  DocumentRef get ref => metadata.selfRef;
 }
